requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
ultralytics==8.0.20
requests==2.28.1
matplotlib==3.5.2
gunicorn==20.1.0
